+++ features/F2/__init__.py
@@ -1,5 +1,11 @@
 """Feature F2 package."""
 
-from . import duplicate_finder, metadata_store, path_links, migrations
+from . import duplicate_finder, metadata_store, path_links, migrations, search_index
 
-__all__ = ["metadata_store", "path_links", "duplicate_finder", "migrations"]
+__all__ = [
+    "metadata_store",
+    "path_links",
+    "duplicate_finder",
+    "migrations",
+    "search_index",
+]
