+++ features/F5/__init__.py
@@ -2,5 +2,6 @@
 
 from . import chunk_module
 from . import chunk_utils
+from . import chunking
 
-__all__ = ["chunk_module", "chunk_utils"]
+__all__ = ["chunk_module", "chunk_utils", "chunking"]
