+++ tests/test_embeddings.py
@@ -1,8 +1,8 @@
 def test_embed_texts_uses_provided_model(monkeypatch):
     import importlib
-    import main as hi
+    import shared.embedding as emb
 
-    importlib.reload(hi)
+    importlib.reload(emb)
 
     class DummyModel:
         class Vec(list):
@@ -10,12 +10,10 @@ def test_embed_texts_uses_provided_model(monkeypatch):
                 return list(self)
 
         def encode(self, texts, convert_to_numpy=True):
-            return [self.Vec([len(t)] * hi.EMBED_DIM) for t in texts]
-
-    import shared.embedding as emb
+            return [self.Vec([len(t)] * emb.EMBED_DIM) for t in texts]
 
     emb.embedding_model = DummyModel()
 
-    result = hi.embed_texts(["a", "bb"])
+    result = emb.embed_texts(["a", "bb"])
 
-    assert result == [[1] * hi.EMBED_DIM, [2] * hi.EMBED_DIM]
+    assert result == [[1] * emb.EMBED_DIM, [2] * emb.EMBED_DIM]
